requirements.txt
click>=8.0.0
fastapi[standard]>=0.109.1
pydantic
python-dotenv
slowapi
starlette>=0.40.0
tiktoken
tomli
uvicorn>=0.11.7
fastapi[all]
jinja2
python-multipart
requests
toml
youtube-transcript-api
yt-dlp
pytube
openai 